requirements.txt
Pillow>=10.0
numpy>=1.24
rembg[cpu]>=2.0
vtracer>=0.6
